refactor: :art: renamed plugin folders
