+++ plugins/blockly-concat-plugin.js
